% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_file.R
\docType{package}
\name{Help}
\alias{Help}
\title{oescolorrrs}
\description{
Takes a ggplot object with colors or fill and applies GSA standard color palettes.
}
\author{
Ben Jaques-Leslie \email{ben.jaques-leslie@gsa.gov}
}
