% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oes_cols.R
\name{oes_cols}
\alias{oes_cols}
\title{Selects specific colors from GSA colors}
\usage{
oes_cols(...)
}
\arguments{
\item{...}{List of colors}
}
\value{
Correct list of colors
}
\description{
Selects specific colors from GSA colors
}
\examples{
oes_cols("blue outer ring","blue inner ring")

}
