% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oes_pal.R
\name{oes_pal}
\alias{oes_pal}
\title{Assigns the selected palette}
\usage{
oes_pal(palette = "primary `accent colors`,", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name of palette to use `primary colors`, `accent colors`, `full colors`,`gradient blue`, `gradient red`, `gradient orange`, `gradient green`, or `gradient magenta`}

\item{reverse}{Boolean indicating whether to reverse the order}

\item{...}{Other parameters passed to ggplot2::discrete_scale() or ggplot2::scale_color_gradientn()}
}
\value{
New color palette based on names
}
\description{
Assigns the selected palette
}
\examples{
oes_pal(palette = "primary colors", reverse = reverse)
}
