% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_oes.R
\name{scale_color_oes}
\alias{scale_color_oes}
\title{GSA colors for color aesthetics in ggplot}
\usage{
scale_color_oes(
  palette = "primary colors",
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{The color palette to use including `primary colors`, `accent colors`, `full colors`, `gradient blue`, `gradient red`, `gradient orange`, `gradient green`, or `gradient magenta`.}

\item{discrete}{Boolean for whether the data is discrete}

\item{reverse}{Boolean to revers the order of the categories}

\item{...}{Other parameters passed to ggplot2::discrete_scale() or ggplot2::scale_color_gradientn()}
}
\value{
ggplot2 object with GSA branding colors
}
\description{
GSA colors for color aesthetics in ggplot
}
\examples{
scale_color_oes(palette = "primary colors", discrete = TRUE, reverse = FALSE,)

}
